requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.14.1
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
kafka-python==2.0.2
transformers==4.35.0
torch==2.1.0
nltk==3.8.1
textblob==0.17.1
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
